Adjust formatting
